Set the action for 'verbose' and 'dryrun' to 'store', use globals as default.
VERBOSE and DRYRUN, as previously set up, aren't used. This allows the
global vars to be set at the top and used as the default when add_argument
is called (and removing 'store_true' because I had to choose between using
the variables or ignoring them. Seems more readable and intuitive.